matching the structure for senior-project
