+++ src/views/HousingLottery/components/PreferenceBox/index.jsx
@@ -1,57 +1,103 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
-  Card,
-  CardContent,
-  CardHeader,
-  Grid,
   Table,
   TableBody,
   TableContainer,
   Typography,
   FormControl,
-  RadioGroup,
   FormControlLabel,
+  Input,
+  Button,
+  RadioGroup,
+  Card,
+  CardContent,
+  CardHeader,
   Radio,
+  Grid,
 } from '@mui/material';
+import housingService from 'services/housing';
 import housing from 'services/housing';
-import styles from '../../../../HousingLottery.module.css';
+import styles from '../../HousingLottery.module.css';
 
-function SurveyQuestionBox() {
-  const [selectedOption, setSelectedOption] = useState(''); // Store the selected option
+const Preference = () => {
+  const [message, setMessage] = useState('');
+  const [morningOrNight, setMorningOrNight] = useState('');
+  const [loudOrQuiet, setLoudOrQuiet] = useState('');
 
-  // Handle the change of the selected option
-  const handleOptionChange = (event) => {
-    setSelectedOption(event.target.value);
+  
+    useEffect(() => {
+     const storedPreferences = localStorage.getItem('userPreferences');
+      if (storedPreferences) {
+        const { morningOrNight, loudOrQuiet } = storedPreferences;
+        setMorningOrNight(morningOrNight || '');
+        setLoudOrQuiet(loudOrQuiet || '');
+      }
+    }, []);
+
+    useEffect(() => {
+      // Save preferences to local storage
+      const storedPreferences = JSON.stringify({ morningOrNight, loudOrQuiet });
+      localStorage.setItem('userPreferences', storedPreferences);
+      }, [morningOrNight, loudOrQuiet]);
+    
+
+  const handleMorningOrNightChange = (event) => {
+    setMorningOrNight(event.target.value);
   };
 
-  // Your survey question text can go here
-  const questionText = 'Are you a night owl or a morning bird?';
+  const handleLoudOrQuietChange = (event) => {
+    setLoudOrQuiet(event.target.value);
+  };
+
+  const handleClick = async () => {
+    // You can access morningOrNight, and loudOrQuiet to submit to your housing service
+    console.log({ morningOrNight, loudOrQuiet});
+    await housingService.addRoommate({morningOrNight, loudOrQuiet});
+  };
 
   return (
-    <Card>
-      <CardHeader title={questionText} />
-      <CardContent>
-        <FormControl component="fieldset">
-          <RadioGroup
-            aria-label="morning-or-night"
-            name="morning-or-night"
-            value={selectedOption}
-            onChange={handleOptionChange}
-          >
-            <FormControlLabel value="night-owl" control={<Radio />} label="Night Owl" />
-            <FormControlLabel value="morning-bird" control={<Radio />} label="Morning Bird" />
-          </RadioGroup>
-          <RadioGroup
-            aria-label="quiet-or-loud"
-            name="quiet-or-loud"
-            value={selectedOption}
-            onChange={handleOptionChange}
-          >
-            <FormControlLabel value="quiet" control={<Radio />} label="Quiet" />
-            <FormControlLabel value="loud" control={<Radio />} label="Loud" />
-          </RadioGroup>
-        </FormControl>
-      </CardContent>
-    </Card>
+<Grid container justifyContent="center">
+      <Grid item xs={12} lg={6} style={{ maxWidth: '500px' }}>
+        <Card>
+          <CardHeader title="Preferences" className={styles.apartment_card_header} />
+          <CardContent>
+            <div>
+              <label>Are you a night owl or a morning bird?</label>
+              <RadioGroup
+                aria-label="morning-or-night"
+                name="morning-or-night"
+                value={morningOrNight}
+                onChange={handleMorningOrNightChange}
+              >
+                <FormControlLabel value="night-owl" control={<Radio />} label="Night Owl" />
+                <FormControlLabel value="morning-bird" control={<Radio />} label="Morning Bird" />
+              </RadioGroup>
+            </div>
+
+            <div>
+              <label>Do you consider yourself quiet or loud in the dorm?</label>
+              <RadioGroup
+                aria-label="loud-or-quiet"
+                name="loud-or-quiet"
+                value={loudOrQuiet}
+                onChange={handleLoudOrQuietChange}
+              >
+                <FormControlLabel value="quiet" control={<Radio />} label="Quiet" />
+                <FormControlLabel value="loud" control={<Radio />} label="Loud" />
+              </RadioGroup>
+            </div>
+          </CardContent>
+          <Grid container justifyContent="flex-end" style={{ padding: '0 16px 16px 0' }}>
+            <Grid item>
+              <Button variant="contained" onClick={handleClick}>
+                Submit
+              </Button>
+            </Grid>
+          </Grid>
+        </Card>
+      </Grid>
+    </Grid>
   );
 }
+
+export default Preference;
